solve issues

+++ pyqit/data/datamodule.py
@@ -67,7 +67,7 @@ def _make_torch_loader(
     class _DS(Dataset):
         def __init__(self, X, y):
             self.X = torch.tensor(X, dtype=torch.float32)
-            self.y = torch.tensor(y, dtype=torch.long)
+            self.y = torch.tensor(y, dtype=torch.float32)
 
         def __len__(self):
             return len(self.X)
